# 2.瑞吉外卖/第二阶段优化/资料/day03/default_OpenAPI.json
{
  "swagger": "2.0",
  "info": {
    "description": "瑞吉外卖接口文档",
    "version": "1.0",
    "title": "瑞吉外卖"
  },
  "host": "172.17.0.60:8200",
  "basePath": "/",
...
  "paths": {
...
    "/dish/{id}": {
      "get": {
        "tags": [
          "菜品管理"
        ],
        "summary": "查询",
        "operationId": "getUsingGET_2",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回值«菜品管理»",
              "originalRef": "返回值«菜品管理»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/setmeal": {
      "post": {
        "tags": [
          "套餐"
        ],
        "summary": "新增",
        "operationId": "saveUsingPOST_6",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "setmeal",
            "description": "setmeal",
            "required": true,
            "schema": {
              "$ref": "#/definitions/套餐1",
              "originalRef": "套餐1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回值«套餐»",
              "originalRef": "返回值«套餐»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "套餐"
        ],
        "summary": "修改",
        "operationId": "updateUsingPUT_6",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "setmeal",
            "description": "setmeal",
            "required": true,
            "schema": {
              "$ref": "#/definitions/套餐1",
              "originalRef": "套餐1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回值«套餐»",
              "originalRef": "返回值«套餐»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "套餐"
        ],
        "summary": "删除",
        "operationId": "deleteUsingDELETE_6",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "ids",
            "required": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回值",
              "originalRef": "返回值"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/setmeal/dish/{id}": {
      "get": {
        "tags": [
          "套餐"
        ],
        "summary": "查询全部菜品",
        "operationId": "getDishUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回值«List«菜品管理»»",
              "originalRef": "返回值«List«菜品管理»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/setmeal/list": {
      "get": {
        "tags": [
          "套餐"
        ],
        "summary": "列表",
        "operationId": "listUsingGET_6",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "query",
            "description": "菜品分类id",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "code",
            "in": "query",
            "description": "编码",
            "required": false,
            "type": "string"
          },
          {
            "name": "createTime",
            "in": "query",
            "description": "创建时间",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "createUser",
            "in": "query",
            "description": "创建人",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "description",
            "in": "query",
            "description": "描述信息",
            "required": false,
            "type": "string"
          },
          {
            "name": "id",
            "in": "query",
            "description": "主键",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "image",
            "in": "query",
            "description": "图片",
            "required": false,
            "type": "string"
          },
          {
            "name": "isDeleted",
            "in": "query",
            "description": "是否删除",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "name",
            "in": "query",
            "description": "套餐名称",
            "required": false,
            "type": "string"
          },
          {
            "name": "price",
            "in": "query",
            "description": "套餐价格",
            "required": false,
            "type": "number"
          },
          {
            "name": "status",
            "in": "query",
            "description": "状态 0:停用 1:启用",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "updateTime",
            "in": "query",
            "description": "更新时间",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "updateUser",
            "in": "query",
            "description": "修改人",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回值«List«套餐»»",
              "originalRef": "返回值«List«套餐»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/setmeal/page": {
      "get": {
        "tags": [
          "套餐"
        ],
        "summary": "分页",
        "operationId": "pageUsingGET_7",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "query",
            "description": "菜品分类id",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "code",
            "in": "query",
            "description": "编码",
            "required": false,
            "type": "string"
          },
          {
            "name": "createTime",
            "in": "query",
            "description": "创建时间",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "createUser",
            "in": "query",
            "description": "创建人",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "description",
            "in": "query",
            "description": "描述信息",
            "required": false,
            "type": "string"
          },
          {
            "name": "id",
            "in": "query",
            "description": "主键",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "image",
            "in": "query",
            "description": "图片",
            "required": false,
            "type": "string"
          },
          {
            "name": "isDeleted",
            "in": "query",
            "description": "是否删除",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "name",
            "in": "query",
            "description": "套餐名称",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": false,
            "type": "integer",
            "default": 1,
            "format": "int32"
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "pageSize",
            "required": false,
            "type": "integer",
            "default": 10,
            "format": "int32"
          },
          {
            "name": "price",
            "in": "query",
            "description": "套餐价格",
            "required": false,
            "type": "number"
          },
          {
            "name": "status",
            "in": "query",
            "description": "状态 0:停用 1:启用",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "updateTime",
            "in": "query",
            "description": "更新时间",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "updateUser",
            "in": "query",
            "description": "修改人",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回值«Page«套餐»»",
              "originalRef": "返回值«Page«套餐»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/setmeal/status/{status}": {
      "post": {
        "tags": [
          "套餐"
        ],
        "summary": "批量起售停售",
        "operationId": "updateUsingPOST_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "id 逗号分隔",
            "required": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "collectionFormat": "multi"
          },
          {
            "name": "status",
            "in": "query",
            "description": "状态：1起售 0停售",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回值",
              "originalRef": "返回值"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/setmeal/{id}": {
      "get": {
        "tags": [
          "套餐"
        ],
        "summary": "查询",
        "operationId": "getUsingGET_6",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回值«套餐»",
              "originalRef": "返回值«套餐»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "OrderItem": {
      "type": "object",
      "properties": {
        "asc": {
          "type": "boolean"
        },
        "column": {
          "type": "string"
        }
      },
      "title": "OrderItem"
    },
...
    "Page«套餐»": {
      "type": "object",
      "properties": {
        "countId": {
          "type": "string"
        },
        "current": {
          "type": "integer",
          "format": "int64"
        },
        "hitCount": {
          "type": "boolean"
        },
        "maxLimit": {
          "type": "integer",
          "format": "int64"
        },
        "optimizeCountSql": {
          "type": "boolean"
        },
        "orders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderItem",
            "originalRef": "OrderItem"
          }
        },
        "pages": {
          "type": "integer",
          "format": "int64"
        },
        "records": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/套餐",
            "originalRef": "套餐"
          }
        },
        "searchCount": {
          "type": "boolean"
        },
        "size": {
          "type": "integer",
          "format": "int64"
        },
        "total": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "Page«套餐»"
    },
...
    "套餐": {
      "type": "object",
      "properties": {
        "categoryId": {
          "type": "integer",
          "format": "int64",
          "description": "菜品分类id"
        },
        "categoryName": {
          "type": "string",
          "description": "菜品分类"
        },
        "code": {
          "type": "string",
          "description": "编码"
        },
        "createTime": {
          "type": "string",
          "format": "date-time",
          "description": "创建时间"
        },
        "createUser": {
          "type": "integer",
          "format": "int64",
          "description": "创建人"
        },
        "description": {
          "type": "string",
          "description": "描述信息"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "主键"
        },
        "image": {
          "type": "string",
          "description": "图片"
        },
        "isDeleted": {
          "type": "integer",
          "format": "int32",
          "description": "是否删除"
        },
        "name": {
          "type": "string",
          "description": "套餐名称"
        },
        "price": {
          "type": "number",
          "description": "套餐价格"
        },
        "setmealDishes": {
          "type": "array",
          "description": "菜品集合",
          "items": {
            "$ref": "#/definitions/套餐菜品关系",
            "originalRef": "套餐菜品关系"
          }
        },
        "status": {
          "type": "integer",
          "format": "int32",
          "description": "状态 0:停用 1:启用"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time",
          "description": "更新时间"
        },
        "updateUser": {
          "type": "integer",
          "format": "int64",
          "description": "修改人"
        }
      },
      "title": "套餐"
    },
    "套餐0": {
      "type": "object",
      "properties": {
        "categoryId": {
          "type": "integer",
          "format": "int64",
          "description": "菜品分类id"
        },
        "code": {
          "type": "string",
          "description": "编码"
        },
        "createTime": {
          "type": "string",
          "format": "date-time",
          "description": "创建时间"
        },
        "createUser": {
          "type": "integer",
          "format": "int64",
          "description": "创建人"
        },
        "description": {
          "type": "string",
          "description": "描述信息"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "主键"
        },
        "image": {
          "type": "string",
          "description": "图片"
        },
        "isDeleted": {
          "type": "integer",
          "format": "int32",
          "description": "是否删除"
        },
        "name": {
          "type": "string",
          "description": "套餐名称"
        },
        "price": {
          "type": "number",
          "description": "套餐价格"
        },
        "status": {
          "type": "integer",
          "format": "int32",
          "description": "状态 0:停用 1:启用"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time",
          "description": "更新时间"
        },
        "updateUser": {
          "type": "integer",
          "format": "int64",
          "description": "修改人"
        }
      },
      "title": "套餐0"
    },
    "套餐1": {
      "type": "object",
      "properties": {
        "categoryId": {
          "type": "integer",
          "format": "int64",
          "description": "菜品分类id"
        },
        "categoryName": {
          "type": "string",
          "description": "菜品分类"
        },
        "code": {
          "type": "string",
          "description": "编码"
        },
        "createTime": {
          "type": "string",
          "format": "date-time",
          "description": "创建时间"
        },
        "createUser": {
          "type": "integer",
          "format": "int64",
          "description": "创建人"
        },
        "description": {
          "type": "string",
          "description": "描述信息"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "主键"
        },
        "image": {
          "type": "string",
          "description": "图片"
        },
        "isDeleted": {
          "type": "integer",
          "format": "int32",
          "description": "是否删除"
        },
        "name": {
          "type": "string",
          "description": "套餐名称"
        },
        "price": {
          "type": "number",
          "description": "套餐价格"
        },
        "setmealDishes": {
          "type": "array",
          "description": "菜品集合",
          "items": {
            "$ref": "#/definitions/套餐菜品关系",
            "originalRef": "套餐菜品关系"
          }
        },
        "status": {
          "type": "integer",
          "format": "int32",
          "description": "状态 0:停用 1:启用"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time",
          "description": "更新时间"
        },
        "updateUser": {
          "type": "integer",
          "format": "int64",
          "description": "修改人"
        }
      },
      "title": "套餐1"
    },
    "套餐菜品关系": {
      "type": "object",
      "properties": {
        "copies": {
          "type": "integer",
          "format": "int32",
          "description": "份数"
        },
        "createTime": {
          "type": "string",
          "format": "date-time",
          "description": "创建时间"
        },
        "createUser": {
          "type": "integer",
          "format": "int64",
          "description": "创建人"
        },
        "dishId": {
          "type": "integer",
          "format": "int64",
          "description": "菜品id"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "主键"
        },
        "isDeleted": {
          "type": "integer",
          "format": "int32",
          "description": "是否删除"
        },
        "name": {
          "type": "string",
          "description": "菜品名称 （冗余字段）"
        },
        "price": {
          "type": "number",
          "description": "菜品原价"
        },
        "setmealId": {
          "type": "integer",
          "format": "int64",
          "description": "套餐id"
        },
        "sort": {
          "type": "integer",
          "format": "int32",
          "description": "排序"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time",
          "description": "更新时间"
        },
        "updateUser": {
          "type": "integer",
          "format": "int64",
          "description": "修改人"
        }
      },
      "title": "套餐菜品关系"
    },
...
    "菜品管理": {
      "type": "object",
      "properties": {
        "categoryId": {
          "type": "integer",
          "format": "int64",
          "description": "菜品分类id"
        },
        "categoryName": {
          "type": "string",
          "description": "菜品分类"
        },
        "code": {
          "type": "string",
          "description": "商品码"
        },
        "copies": {
          "type": "integer",
          "format": "int32",
          "description": "份数"
        },
        "createTime": {
          "type": "string",
          "format": "date-time",
          "description": "创建时间"
        },
        "createUser": {
          "type": "integer",
          "format": "int64",
          "description": "创建人"
        },
        "description": {
          "type": "string",
          "description": "描述信息"
        },
        "flavors": {
          "type": "array",
          "description": "口味集合",
          "items": {
            "$ref": "#/definitions/菜品口味关系表",
            "originalRef": "菜品口味关系表"
          }
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "主键"
        },
        "image": {
          "type": "string",
          "description": "图片"
        },
        "isDeleted": {
          "type": "integer",
          "format": "int32",
          "description": "是否删除"
        },
        "name": {
          "type": "string",
          "description": "菜品名称"
        },
        "price": {
          "type": "number",
          "description": "菜品价格"
        },
        "sort": {
          "type": "integer",
          "format": "int32",
          "description": "顺序"
        },
        "status": {
          "type": "integer",
          "format": "int32",
          "description": "0 停售 1 起售"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time",
          "description": "更新时间"
        },
        "updateUser": {
          "type": "integer",
          "format": "int64",
          "description": "修改人"
        }
      },
      "title": "菜品管理"
    },
...
    "返回值": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "编码：1成功，0和其它数字为失败"
        },
        "data": {
          "type": "object",
          "description": "返回对象"
        },
        "map": {
          "type": "object",
          "description": "动态字段"
        },
        "msg": {
          "type": "string",
          "description": "返回信息"
        }
      },
      "title": "返回值"
    },
...
    "返回值«List«套餐»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "编码：1成功，0和其它数字为失败"
        },
        "data": {
          "type": "array",
          "description": "返回对象",
          "items": {
            "$ref": "#/definitions/套餐0",
            "originalRef": "套餐0"
          }
        },
        "map": {
          "type": "object",
          "description": "动态字段",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string",
          "description": "返回信息"
        }
      },
      "title": "返回值«List«套餐»»"
    },
...
    "返回值«List«菜品管理»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "编码：1成功，0和其它数字为失败"
        },
        "data": {
          "type": "array",
          "description": "返回对象",
          "items": {
            "$ref": "#/definitions/菜品管理",
            "originalRef": "菜品管理"
          }
        },
        "map": {
          "type": "object",
          "description": "动态字段",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string",
          "description": "返回信息"
        }
      },
      "title": "返回值«List«菜品管理»»"
    },
...
    "返回值«Page«套餐»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "编码：1成功，0和其它数字为失败"
        },
        "data": {
          "description": "返回对象",
          "$ref": "#/definitions/Page«套餐»",
          "originalRef": "Page«套餐»"
        },
        "map": {
          "type": "object",
          "description": "动态字段",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string",
          "description": "返回信息"
        }
      },
      "title": "返回值«Page«套餐»»"
    },
...
    "返回值«套餐»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "编码：1成功，0和其它数字为失败"
        },
        "data": {
          "description": "返回对象",
          "$ref": "#/definitions/套餐",
          "originalRef": "套餐"
        },
        "map": {
          "type": "object",
          "description": "动态字段",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string",
          "description": "返回信息"
        }
      },
      "title": "返回值«套餐»"
    },
...
    "返回值«菜品管理»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "编码：1成功，0和其它数字为失败"
        },
        "data": {
          "description": "返回对象",
          "$ref": "#/definitions/菜品管理",
          "originalRef": "菜品管理"
        },
        "map": {
          "type": "object",
          "description": "动态字段",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string",
          "description": "返回信息"
        }
      },
      "title": "返回值«菜品管理»"
    },
...
  }
}